add update

+++ src/belajar4/kocheng.php
@@ -37,8 +37,8 @@
         <tr align="center">
             <td><?= $i; ?></td>
             <td>
-                <a href="">Ubah</a> |
-                <a href="hapus.php?id=<?= $row["id"]; ?>">Hapus</a>
+                <a href="ubah.php?id=<?= $row["id"]; ?>">Ubah</a> |
+                <a href="hapus.php?id=<?= $row["id"]; ?>" onclick="return confirm('Yakin ingin menghapus kocheng ini??')">Hapus</a>
             </td>
             <td><img src="img/<?= $row["gambar"]; ?>" width="50"></td>
             <td><?= $row["nama"]; ?></td>
